Refactor StackState repo creation to charts package.
Moved the `CreateClusterRepo` function from the observability package to the charts package for better organizational clarity. Simplified test cleanup logic and removed unused imports and code for improved maintainability and readability.
